Symmetric stateless reset
This is disgusting.  It's still not symmetric, because only the server
can propose new connection IDs.  That means that the best option is to
define a new message for providing the client token.

The client can't send a token in transport parameters because those
aren't encrypted.  That means that it needs to send a frame.  I
considered using NEW_CONNECTION_ID, which would also allow for
asymmetric connection IDs, but that doesn't work very well for other
reasons: it makes the stateless reset trivially distinguishable for one.

Thus, we have this.

I would prefer NOT to merge this.  It's complicated and gross and I
still believe we don't need to do this.  The primary reason, which I
don't think was correctly articulated at the interim, was that this
feature (stateless reset) is only really useful if you have are
stationary and therefore likely to receive packets immediately after a
reboot or crash.  Most endpoints, particularly with IPv6, will have new
IP addresses in that state and will not be listening on ports.  I
realize that means for stationary clients that means that they will need
to fall back to ICMP, or the server detecting a broken path, but I think
that is preferable to this mess.